Move StartTLS functionality into a subpackage, enabling use elsewhere
This moves two functions out of main.go.
The only change to the functions was to make them take arguments, instead of using global variables in main.

The mysql and psql imports are moved here too.
